fix(TransformableImage): prevent state updates if the component has been unmounted
This now happens a lot more since we’re now using a virtualized list,
which means some images might be unmounted when Image.getSize has
finished.
This does not prevent crashes per se, but it’s a good practice with
react.

Also removed console.log in development that slowed the hell out of the
gallery.