Add missing bind methods for gesture responder.
Somehow the context is getting dropped *some of the time*, so that
`this` becomes undefined inside `onResponderGrant`. Binding the
methods to the instance fixes the issue.

It's unclear how it's only happening some of the time -- usually
binding issues will become immediate apparent with React -- so my
best guess is there's some detail inside the React Native system
where the context can be dropped.